workers/constructors/Worker/terminate.html

<!--
while (true)
  postMessage(1);
/*
-->
<!doctype html>
<title>terminate()</title>
<link rel=help href="http://www.whatwg.org/html/#dom-worker-terminate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../workers.js"></script>
<div id="log"></div>
<script>
var t = async_test();
t.step(function() {
  var worker = new Worker('#');
  var i = 0;
  worker.onmessage = function(e) {
    i++;
    this.terminate();
  }
  setTimeout(function() {
    t.step(function() {
      assert_equals(i, 1);
    });
    t.done();
  }, 200);
});
</script>
<!--
*/
//-->
